requirements.txt
streamlit>=1.44.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0
pillow>=10.0.0
pandas>=2.2.0
plotly>=5.24.0
pydantic>=2.8.0
numpy>=1.26.0
